to help with review of code
solveMuscleRedundancy.m now has all the OCPs defined in it. This should help with the review of code since the old version had the same format. The OCPs are directly copied from the functions where I initially wrote them.
solveMuscleRedundancy_OCPinFunctions.m has all the OCPs in separate files. My vision is that this should be implemented. So once solveMuscleRedundancy.m has been reviewed, we can remove it and replace it by solveMuscleRedundancy_OCPinFunctions.m

If this method does not actually help with the review, this commit can simply be reversed.