Backward compatibility addressed
This update has multiple minor changes that essentially address backward compatibility. Now all previous examples work. I also tested that the results of the old code (after the corrections I mentioned earlier) are the same as the results of the new code.
Additionally, I added an example where trials with different number of DOFs are used in the same optimization.
I think that the coding part is now complete.

Things left on To Do list:
1. I added some warnings. I will look at some of the warnings and check if they are really required. I will at least make them displayed in a more easy to read manner.
2. Add comments and headers to all files.
3. Update the readme